requirements.txt
streamlit
PyMuPDF
pdfplumber
python-docx
Pillow
pytesseract
fpdf
scikit-learn